requirements.txt
pymongo[srv]==4.2.0
pymongo
numpy
pandas
scikit-learn
python-dotenv
PyAML
dill
scipy
certifi
imblearn==0.0
xgboost==1.6.2

-e .              # editable mode